- Refactor mapper and models - Add Glide
