+++ app/build.gradle
@@ -66,7 +66,7 @@ dependencies {
     kapt("androidx.room:room-compiler:$room_version")
     implementation("androidx.room:room-ktx:$room_version")
 
-    implementation 'com.squareup.picasso:picasso:2.71828'
+    implementation 'com.github.bumptech.glide:glide:4.12.0'
 
     implementation 'androidx.work:work-runtime-ktx:2.7.1'
 }
