+++ app/src/main/java/ua/com/koval/andrey/cryptoapp/data/database/CoinInfoDbModel.kt
@@ -13,5 +13,5 @@ data class CoinInfoDbModel(
     val highDay: String?,
     val lowDay: String?,
     val lastMarket: String?,
-    val imageUrl: String?
+    val imageUrl: String
 )
