+++ app/src/main/java/ua/com/koval/andrey/cryptoapp/data/network/ApiFactory.kt
@@ -5,8 +5,7 @@ import retrofit2.converter.gson.GsonConverterFactory
 
 object ApiFactory {
 
-    private const val BASE_URL : String = "https://min-api.cryptocompare.com/data/"
-    const val BASE_IMG_URL : String = "https://cryptocompare.com"
+    private const val BASE_URL: String = "https://min-api.cryptocompare.com/data/"
 
     private val retrofit = Retrofit.Builder()
         .addConverterFactory(GsonConverterFactory.create())
