+++ app/src/main/java/ua/com/koval/andrey/cryptoapp/domain/CoinInfo.kt
@@ -5,9 +5,9 @@ data class CoinInfo(
     val fromSymbol: String,
     val toSymbol: String?,
     val price: String?,
-    val lastUpdate: Long?,
+    val lastUpdate: String,
     val highDay: String?,
     val lowDay: String?,
     val lastMarket: String?,
-    val imageUrl: String?
+    val imageUrl: String
 )
