--- app/src/main/java/ua/com/koval/andrey/cryptoapp/utils/Times.kt
@@ -1,15 +1,0 @@
-package ua.com.koval.andrey.cryptoapp.utils
-
-import java.sql.Timestamp
-import java.text.SimpleDateFormat
-import java.util.*
-
-fun convertTimestampToTime(timestamp: Long?): String{
-    if (timestamp == null) return ""
-    val stamp = Timestamp(timestamp * 1000)
-    val date = Date(stamp.time)
-    val pattern = "HH:mm:ss"
-    val sdf = SimpleDateFormat(pattern, Locale.getDefault())
-    sdf.timeZone = TimeZone.getDefault()
-    return sdf.format(date)
-}
